created component MeasureOverview
