+++ pages/measures/[id].jsx
@@ -3,12 +3,8 @@ import useSWR from 'swr';
 import Link from 'next/link';
 import Image from 'next/image';
 import Layout from '../../components/layouts/layout';
-import WindmillIcon from '../../public/winturbines.svg';
-import IcontWood from '../../public/icons/houtbouwIconBg.svg';
-import RTooltip from '../../components/r-ladder-tooltip';
-import JHTooltip from '../../components/juridische-houdbaarheid-tooltip';
-import JITooltip from '../../components/juridische-invloed-tooltip';
 import LinkIcon from '../../components/link-icon';
+import MeasureOverview from '../../components/measure-overview';
 
 const formatDate = (date) => {
   let dateObject = new Date(date);
@@ -55,9 +51,7 @@ const fetcher = async (url) => {
   return data;
 };
 
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
+
 
 export default function Law() {
   const { query } = useRouter();
@@ -66,7 +60,7 @@ export default function Law() {
 
   if (error) return <div>{error.message} </div>;
   if (!data) return <div>Loading...</div>;
-
+  console.log(data.casus)
   return (
     <Layout>
       <div className='gradient-bg'>
@@ -80,135 +74,12 @@ export default function Law() {
               <a className='text-greenLink breadcrumb'>← Terug</a>
             </Link>
           )}
+
           {/* TODO: refactor this and have same component at top and bottom */}
-          <div className='block lg:hidden'>
-            <h1 className='my-9 text-green1 mobile sm:main'>{data.titel}</h1>
-            <div className='container pb-2'>
-              <div className='container-image'>
-                {data.casus === 'Houtbouw' ? (
-                  <div className='container-image'>
-                    <Image src={IcontWood} alt='Icon of a Wood Log' />
-                  </div>
-                ) : (
-                  <div className='container-image'>
-                    <Image src={WindmillIcon} alt='Icon of a Wood Log' />
-                  </div>
-                )}
-              </div>
-              <div className=''>
-                <Link href={'/' + data.casus.replace(/\s+/g, '-').toLowerCase()}>
-                  <a>
-                    <span className='font-openSans font-bold pl-2 text-greenLink'>
-                      {data.casus}
-                    </span>
-                  </a>
-                </Link>
-              </div>
-            </div>
+          {/* start here */}
+          <MeasureOverview data = {data} viewport = 'desktop'/>
+          {/* finish here */}
 
-            <div className='py-5 border-t-2 border-grey2 '>
-              <div className='flex pb-2'>
-                <span className='font-manrope font-semibold text-lg text-black1'>R-ladder</span>
-                <RTooltip>
-                  <svg className='w-6 h-6 fill-current text-black mx-2' viewBox='0 0 26 26'>
-                    <circle cx='12' cy='15' r='10' fill='#979797' />
-                    <path
-                      d='M10.7031 10.0078C10.7031 9.23177 11.1354 8.84375 12 8.84375C12.8646 8.84375 13.2969 9.23177 13.2969 10.0078C13.2969 10.3776 13.1875 10.6667 12.9688 10.875C12.7552 11.0781 12.4323 11.1797 12 11.1797C11.1354 11.1797 10.7031 10.7891 10.7031 10.0078ZM13.1875 21H10.8047V12.2656H13.1875V21Z'
-                      fill='black'
-                    />
-                  </svg>
-                </RTooltip>
-              </div>
-
-              <span className='block-inline font-semibold text-base text-gray-900'>
-                {data.R1 && <span className='bg-green2 text-white rounded-full p-1 mr-2'>R1</span>}
-                {data.R2 && <span className='bg-green2 text-white rounded-full p-1 mr-2'>R2</span>}
-                {data.R3 && <span className='bg-green2 text-white rounded-full p-1 mr-2'>R3</span>}
-                {data.R4 && <span className='bg-green2 text-white rounded-full p-1 mr-2'>R4</span>}
-                {data.R5 && <span className='bg-green2 text-white rounded-full p-1 mr-2'>R5</span>}
-                {data.R6 && <span className='bg-green2 text-white rounded-full p-1 mr-2'>R6</span>}
-              </span>
-            </div>
-
-            <div className='py-5'>
-              <div className='relative border-t-2 border-grey2 pt-4'>
-                <div className='font-manrope font-semibold text-lg text-black1 pb-2'>
-                  Subrechtsgebied
-                </div>
-              </div>
-
-              <div className='font-manrope font-normal text-base'>
-                <p>{data.subrechtsgebied}</p>
-              </div>
-            </div>
-
-            <div className='py-5'>
-              <div className='relative flex justify-between border-t-2 border-grey2 pt-2'>
-                <div className='flex pb-2'>
-                  <span className='font-manrope font-semibold text-lg text-black1'>
-                    Juridisch invloed
-                  </span>
-                  <JITooltip>
-                    <svg className='w-6 h-6 fill-current text-black mx-2' viewBox='0 0 26 26'>
-                      <circle cx='12' cy='15' r='10' fill='#979797' />
-                      <path
-                        d='M10.7031 10.0078C10.7031 9.23177 11.1354 8.84375 12 8.84375C12.8646 8.84375 13.2969 9.23177 13.2969 10.0078C13.2969 10.3776 13.1875 10.6667 12.9688 10.875C12.7552 11.0781 12.4323 11.1797 12 11.1797C11.1354 11.1797 10.7031 10.7891 10.7031 10.0078ZM13.1875 21H10.8047V12.2656H13.1875V21Z'
-                        fill='black'
-                      />
-                    </svg>
-                  </JITooltip>
-                </div>
-              </div>
-
-              <div className='mt-3 flex items-center'>
-                <span className='pr-5 font-manrope font-normal text-base'>LAAG</span>
-                {[0, 1, 2, 3, 4].map((rating) => (
-                  <div
-                    key={rating}
-                    className={classNames(
-                      data.juridische_invloed > rating ? 'score-true' : 'score-false',
-                      'mr-5 h-6 w-6 flex-shrink-0 rounded-full',
-                    )}
-                    aria-hidden='true'
-                  />
-                ))}
-                HOOG
-              </div>
-            </div>
-
-            <div className='py-5'>
-              <div className='relative flex justify-between border-t-2 border-grey2 pt-2'>
-                <div className='flex pb-2'>
-                  <span className='font-manrope font-semibold text-lg text-black1 pb-2'>
-                    Juridisch houdbaarheid
-                  </span>
-                  <JHTooltip>
-                    <svg className='w-6 h-6 fill-current text-black mx-2' viewBox='0 0 26 26'>
-                      <circle cx='12' cy='15' r='10' fill='#979797' />
-                      <path
-                        d='M10.7031 10.0078C10.7031 9.23177 11.1354 8.84375 12 8.84375C12.8646 8.84375 13.2969 9.23177 13.2969 10.0078C13.2969 10.3776 13.1875 10.6667 12.9688 10.875C12.7552 11.0781 12.4323 11.1797 12 11.1797C11.1354 11.1797 10.7031 10.7891 10.7031 10.0078ZM13.1875 21H10.8047V12.2656H13.1875V21Z'
-                        fill='black'
-                      />
-                    </svg>
-                  </JHTooltip>
-                </div>
-              </div>
-              <div className='mt-3 flex items-center'>
-                <span className='pr-5 font-manrope font-normal text-base'> LAAG</span>
-                {[0, 1, 2, 3, 4].map((rating) => (
-                  <div
-                    key={rating}
-                    className={classNames(
-                      data.juridische_houdbaarheid > rating ? 'score-true' : 'score-false',
-                      'mr-5 h-6 w-6 flex-shrink-0 rounded-full',
-                    )}
-                    aria-hidden='true'
-                  />
-                ))}
-                HOOG
-              </div>
-            </div>
-          </div>
           <div className='grid grid-cols-1 sm:grid-cols-3 '>
             <div className='w-full sm:max-w-3xl pb-20 col-span-2 '>
               <h1 className='hidden lg:block my-9 text-green1 mobile sm:main'>{data.titel}</h1>
@@ -270,6 +141,8 @@ export default function Law() {
                 <h2 className='pb-2 mobile sm:main'>{data.kop_5_juridische_toelichting}</h2>
                 <p className='body-text-mobile sm:body-text'>{data.juridische_toelichting}</p>
               </div>
+                
+                {/* refactor table to have different design on mobile */}
               <table className='table-fixed w-full mt-5'>
                 <tbody>
                   <tr className='my-10 border-b-2 border-t-2'>
@@ -332,143 +205,13 @@ export default function Law() {
                 </tbody>
               </table>
             </div>
-            <div className='hidden lg:block float-right'>
-              <div className='container pb-2'>
-                {data.casus === 'Houtbouw' ? (
-                  <div className='container-image'>
-                    <Image src={IcontWood} alt='Icon of a Wood Log' />
-                  </div>
-                ) : (
-                  <div className='container-image'>
-                    <Image src={WindmillIcon} alt='Icon of a Wood Log' />
-                  </div>
-                )}
-                <div className=''>
-                  <Link href={'/' + data.casus.replace(/\s+/g, '-').toLowerCase()}>
-                    <a>
-                      <span className='font-openSans font-bold pl-2 text-greenLink'>
-                        {data.casus}
-                      </span>
-                    </a>
-                  </Link>
-                </div>
-              </div>
+              
 
-              <div className='py-5 border-t-2 border-grey2 '>
-                <div className='flex pb-2'>
-                  <span className='font-manrope font-semibold text-lg text-black1'>R-ladder</span>
-                  <RTooltip>
-                    <svg className='w-6 h-6 fill-current text-black mx-2' viewBox='0 0 26 26'>
-                      <circle cx='12' cy='15' r='10' fill='#979797' />
-                      <path
-                        d='M10.7031 10.0078C10.7031 9.23177 11.1354 8.84375 12 8.84375C12.8646 8.84375 13.2969 9.23177 13.2969 10.0078C13.2969 10.3776 13.1875 10.6667 12.9688 10.875C12.7552 11.0781 12.4323 11.1797 12 11.1797C11.1354 11.1797 10.7031 10.7891 10.7031 10.0078ZM13.1875 21H10.8047V12.2656H13.1875V21Z'
-                        fill='black'
-                      />
-                    </svg>
-                  </RTooltip>
-                </div>
+              {/* start here */}
+            <MeasureOverview data={data} viewport ='mobile'/>
+                {/* finish here */}
 
-                <span className='block-inline font-semibold text-base text-gray-900'>
-                  {data.R1 && (
-                    <span className='bg-green2 text-white rounded-full p-1 mr-2'>R1</span>
-                  )}
-                  {data.R2 && (
-                    <span className='bg-green2 text-white rounded-full p-1 mr-2'>R2</span>
-                  )}
-                  {data.R3 && (
-                    <span className='bg-green2 text-white rounded-full p-1 mr-2'>R3</span>
-                  )}
-                  {data.R4 && (
-                    <span className='bg-green2 text-white rounded-full p-1 mr-2'>R4</span>
-                  )}
-                  {data.R5 && (
-                    <span className='bg-green2 text-white rounded-full p-1 mr-2'>R5</span>
-                  )}
-                  {data.R6 && (
-                    <span className='bg-green2 text-white rounded-full p-1 mr-2'>R6</span>
-                  )}
-                </span>
-              </div>
 
-              <div className='py-5'>
-                <div className='relative border-t-2 border-grey2 pt-4'>
-                  <div className='font-manrope font-semibold text-lg text-black1 pb-2'>
-                    Subrechtsgebied
-                  </div>
-                </div>
-
-                <div className='font-manrope font-normal text-base'>
-                  <p>{data.subrechtsgebied}</p>
-                </div>
-              </div>
-
-              <div className='py-5'>
-                <div className='relative flex justify-between border-t-2 border-grey2 pt-2'>
-                  <div className='flex pb-2'>
-                    <span className='font-manrope font-semibold text-lg text-black1'>
-                      Juridisch invloed
-                    </span>
-                    <JITooltip>
-                      <svg className='w-6 h-6 fill-current text-black mx-2' viewBox='0 0 26 26'>
-                        <circle cx='12' cy='15' r='10' fill='#979797' />
-                        <path
-                          d='M10.7031 10.0078C10.7031 9.23177 11.1354 8.84375 12 8.84375C12.8646 8.84375 13.2969 9.23177 13.2969 10.0078C13.2969 10.3776 13.1875 10.6667 12.9688 10.875C12.7552 11.0781 12.4323 11.1797 12 11.1797C11.1354 11.1797 10.7031 10.7891 10.7031 10.0078ZM13.1875 21H10.8047V12.2656H13.1875V21Z'
-                          fill='black'
-                        />
-                      </svg>
-                    </JITooltip>
-                  </div>
-                </div>
-
-                <div className='mt-3 flex items-center'>
-                  <span className='pr-5 font-manrope font-normal text-base'>LAAG</span>
-                  {[0, 1, 2, 3, 4].map((rating) => (
-                    <div
-                      key={rating}
-                      className={classNames(
-                        data.juridische_invloed > rating ? 'score-true' : 'score-false',
-                        'mr-5 h-6 w-6 flex-shrink-0 rounded-full',
-                      )}
-                      aria-hidden='true'
-                    />
-                  ))}
-                  HOOG
-                </div>
-              </div>
-
-              <div className='py-5'>
-                <div className='relative flex justify-between border-t-2 border-grey2 pt-2'>
-                  <div className='flex pb-2'>
-                    <span className='font-manrope font-semibold text-lg text-black1 '>
-                      Juridisch houdbaarheid
-                    </span>
-                    <JHTooltip>
-                      <svg className='w-6 h-6 fill-current text-black mx-2' viewBox='0 0 26 26'>
-                        <circle cx='12' cy='15' r='10' fill='#979797' />
-                        <path
-                          d='M10.7031 10.0078C10.7031 9.23177 11.1354 8.84375 12 8.84375C12.8646 8.84375 13.2969 9.23177 13.2969 10.0078C13.2969 10.3776 13.1875 10.6667 12.9688 10.875C12.7552 11.0781 12.4323 11.1797 12 11.1797C11.1354 11.1797 10.7031 10.7891 10.7031 10.0078ZM13.1875 21H10.8047V12.2656H13.1875V21Z'
-                          fill='black'
-                        />
-                      </svg>
-                    </JHTooltip>
-                  </div>
-                </div>
-                <div className='mt-3 flex items-center'>
-                  <span className='pr-5 font-manrope font-normal text-base'> LAAG</span>
-                  {[0, 1, 2, 3, 4].map((rating) => (
-                    <div
-                      key={rating}
-                      className={classNames(
-                        data.juridische_houdbaarheid > rating ? 'score-true' : 'score-false',
-                        'mr-5 h-6 w-6 flex-shrink-0 rounded-full',
-                      )}
-                      aria-hidden='true'
-                    />
-                  ))}
-                  HOOG
-                </div>
-              </div>
-            </div>
           </div>
         </div>
       </div>
